feat: destructured page & separate client side
